backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Audio processing
sounddevice==0.4.6
scipy>=1.10.0
numpy>=1.21.0

# OpenAI
openai==1.3.7

# Environment
python-dotenv==1.0.0

# File handling
python-multipart==0.0.6

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
